+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 All notable changes to Porsche backend project will be documented in this file.
 
 ## Unreleased
+
+##[0.1.1]() - 2021-05-19
 ### Added
-- Users can find, create and delete chat rooms
+- Users can find, create and delete chat lobbies
+- Users can find, update and delete lobby messages
 
 ##[0.1.0]() - 2021-05-19
 ### Added